feat(quiz): add totalPoints, pointsDistribution, and clearer validation

+++ be/src/main/java/swdchatbox/modules/quiz/dto/response/QuizAnswerResultResponse.java
@@ -12,6 +12,6 @@ public class QuizAnswerResultResponse {
     private QuestionTypeResponse questionType;
     private String questionText;
     private Boolean isCorrect;
-    private Integer scoreEarned;
-    private Integer maxScore;
+    private Double scoreEarned;
+    private Double maxScore;
 }
