+++ be/src/main/java/swdchatbox/modules/quiz/dto/response/QuizResponse.java
@@ -21,7 +21,7 @@ public class QuizResponse {
     private String description;
     private QuizStatus status;
     private Integer timeLimitMinutes;
-    private Integer totalPoints;
+    private Double totalPoints;
     private Integer questionCount;
     private Boolean active;
     private Boolean aiGenerated;
